radically simplify the HTTP/2 test
Previously, the HTTP/2 test used the Jetty HTTP/2 client library
to implement a minimal JAX-RS client connector that would support
HTTP/2. It didn't properly support e.g. a `HEAD` HTTP method.
With this commit, the test now directly uses the Jetty HTTP client
high-level library, together with the HTTP/2 transport library.
This makes the test code much shorter (the entire JAX-RS client
connector is removed).

I've decided to do this because the Akamai HTTP/2 test site
no longer returns the `myproto` header in the response to
a `GET` request. It still returns the header in the response
to a `HEAD` request, but if we need to rewrite stuff (because
`HEAD` doesn't work with the current JAX-RS client and it isn't
trivial to fix this), we can as well just inspect the status
line of the response, which indicates the protocol directly.
That way, we no longer have to rely on a nonstandard header,
which can disappear at any time.